delete fx and from go.mod
