+++ go.mod
@@ -15,7 +15,6 @@ require (
 	github.com/spf13/viper v1.18.2
 	github.com/stretchr/testify v1.8.4
 	github.com/wI2L/jsondiff v0.5.1
-	go.uber.org/fx v1.20.1
 	go.uber.org/mock v0.4.0
 	golang.org/x/mod v0.13.0
 )
@@ -70,9 +69,11 @@ require (
 	go.opentelemetry.io/otel/sdk v1.22.0 // indirect
 	go.opentelemetry.io/otel/trace v1.22.0 // indirect
 	go.uber.org/atomic v1.9.0 // indirect
+<<<<<<< HEAD
 	go.uber.org/dig v1.17.0 // indirect
+=======
+>>>>>>> 895f81e (delete fx and from go.mod)
 	go.uber.org/multierr v1.9.0 // indirect
-	go.uber.org/zap v1.23.0 // indirect
 	golang.org/x/exp v0.0.0-20230905200255-921286631fa9 // indirect
 	golang.org/x/sys v0.16.0 // indirect
 	golang.org/x/text v0.14.0 // indirect
