+++ internal/node/api/router.go
@@ -7,7 +7,6 @@ import (
 	"github.com/eagraf/habitat-new/internal/node/controller"
 	"github.com/gorilla/mux"
 	"github.com/rs/zerolog"
-	"go.uber.org/fx"
 )
 
 type Route interface {
@@ -16,16 +15,6 @@ type Route interface {
 	// Pattern reports the path at which this is registered.
 	Pattern() string
 	Method() string
-}
-
-// AsRoute annotates the given constructor to state that
-// it provides a route to the "routes" group.
-func AsRoute(f any) any {
-	return fx.Annotate(
-		f,
-		fx.As(new(Route)),
-		fx.ResultTags(`group:"routes"`),
-	)
 }
 
 func NewRouter(routes []Route, logger *zerolog.Logger, nodeController controller.NodeController, nodeConfig *config.NodeConfig) *mux.Router {
