--- internal/node/config/fx.go
@@ -1,9 +1,0 @@
-package config
-
-func NewNodeConfig() (*NodeConfig, error) {
-	err := loadEnv()
-	if err != nil {
-		return nil, err
-	}
-	return loadConfig()
-}
