+++ internal/node/logging/logger.go
@@ -1,11 +1,8 @@
 package logging
 
 import (
-	"strings"
-
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
-	"go.uber.org/fx/fxevent"
 )
 
 func NewLogger() *zerolog.Logger {
@@ -13,96 +10,4 @@ func NewLogger() *zerolog.Logger {
 	logger := zerolog.New(out).With().Timestamp().Logger()
 	log.Logger = logger
 	return &logger
-}
-
-// Implement the fxevent.Logger interface
-// Adapted from https://github.com/ipfans/fxlogger
-type FxEventLoggerWrapper struct {
-	Logger *zerolog.Logger
-}
-
-func (l *FxEventLoggerWrapper) LogEvent(event fxevent.Event) {
-	switch e := event.(type) {
-	case *fxevent.OnStartExecuting:
-		l.Logger.Info().Str("callee", e.FunctionName).
-			Str("caller", e.CallerName).
-			Msg("OnStart hook executing")
-	case *fxevent.OnStartExecuted:
-		if e.Err != nil {
-			l.Logger.Warn().Err(e.Err).
-				Str("callee", e.FunctionName).
-				Str("caller", e.CallerName).
-				Msg("OnStart hook failed")
-		} else {
-			l.Logger.Info().Str("callee", e.FunctionName).
-				Str("caller", e.CallerName).
-				Str("runtime", e.Runtime.String()).
-				Msg("OnStart hook executed")
-		}
-	case *fxevent.OnStopExecuting:
-		l.Logger.Info().Str("callee", e.FunctionName).
-			Str("caller", e.CallerName).
-			Msg("OnStop hook executing")
-	case *fxevent.OnStopExecuted:
-		if e.Err != nil {
-			l.Logger.Warn().Err(e.Err).
-				Str("callee", e.FunctionName).
-				Str("callee", e.CallerName).
-				Msg("OnStop hook failed")
-		} else {
-			l.Logger.Info().Str("callee", e.FunctionName).
-				Str("caller", e.CallerName).
-				Str("runtime", e.Runtime.String()).
-				Msg("OnStop hook executed")
-		}
-	case *fxevent.Supplied:
-		if e.Err != nil {
-			l.Logger.Warn().Err(e.Err).Str("type", e.TypeName).Msg("supplied")
-		} else {
-			l.Logger.Info().Str("type", e.TypeName).Msg("supplied")
-		}
-	case *fxevent.Provided:
-		for _, rtype := range e.OutputTypeNames {
-			l.Logger.Info().Str("type", rtype).
-				Str("constructor", e.ConstructorName).
-				Msg("provided")
-		}
-		if e.Err != nil {
-			l.Logger.Error().Err(e.Err).Msg("error encountered while applying options")
-		}
-	case *fxevent.Invoking:
-		// Do nothing. Will log on Invoked.
-
-	case *fxevent.Invoked:
-		if e.Err != nil {
-			l.Logger.Error().Err(e.Err).Str("stack", e.Trace).
-				Str("function", e.FunctionName).Msg("invoke failed")
-		} else {
-			l.Logger.Info().Str("function", e.FunctionName).Msg("invoked")
-		}
-	case *fxevent.Stopping:
-		l.Logger.Info().Str("signal", strings.ToUpper(e.Signal.String())).Msg("received signal")
-	case *fxevent.Stopped:
-		if e.Err != nil {
-			l.Logger.Error().Err(e.Err).Msg("stop failed")
-		}
-	case *fxevent.RollingBack:
-		l.Logger.Error().Err(e.StartErr).Msg("start failed, rolling back")
-	case *fxevent.RolledBack:
-		if e.Err != nil {
-			l.Logger.Error().Err(e.Err).Msg("rollback failed")
-		}
-	case *fxevent.Started:
-		if e.Err != nil {
-			l.Logger.Error().Err(e.Err).Msg("start failed")
-		} else {
-			l.Logger.Info().Msg("started")
-		}
-	case *fxevent.LoggerInitialized:
-		if e.Err != nil {
-			l.Logger.Error().Err(e.Err).Msg("custom logger initialization failed")
-		} else {
-			l.Logger.Info().Str("function", e.ConstructorName).Msg("initialized custom fxevent.Logger")
-		}
-	}
 }
