+++ internal/node/processes/manager.go
@@ -4,6 +4,8 @@ import (
 	"fmt"
 
 	"github.com/eagraf/habitat-new/core/state/node"
+	"github.com/eagraf/habitat-new/internal/node/controller"
+	"github.com/eagraf/habitat-new/internal/node/hdb"
 )
 
 type RunningProcess struct {
@@ -34,6 +36,31 @@ func newBaseProcessManager() *BaseProcessManager {
 		processDrivers: make(map[string]ProcessDriver),
 		processes:      make(map[string]*RunningProcess),
 	}
+}
+
+func NewProcessManager(drivers []ProcessDriver) ProcessManager {
+	pm := newBaseProcessManager()
+	for _, driver := range drivers {
+		pm.processDrivers[driver.Type()] = driver
+	}
+	return pm
+}
+
+func NewProcessManagerStateUpdateSubscriber(processManager ProcessManager, controller controller.NodeController) (*hdb.IdempotentStateUpdateSubscriber, error) {
+	return hdb.NewIdempotentStateUpdateSubscriber(
+		"StartProcessSubscriber",
+		node.SchemaName,
+		[]hdb.IdempotentStateUpdateExecutor{
+			&StartProcessExecutor{
+				processManager: processManager,
+				nodeController: controller,
+			},
+		},
+		&ProcessRestorer{
+			processManager: processManager,
+			nodeController: controller,
+		},
+	)
 }
 
 func (pm *BaseProcessManager) ListProcesses() ([]*node.ProcessState, error) {
